renamed package and fixed compile errors
