+++ src/org/eclipse/jdt/neon/formatter/DefaultCodeFormatterOptions.java
@@ -14,7 +14,7 @@
  *     Lars Vogel <Lars.Vogel@vogella.com> - Contributions for
  *     						Bug 473178
  *******************************************************************************/
-package org.eclipse.jdt.internal.formatter;
+package org.eclipse.jdt.neon.formatter;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -22,6 +22,7 @@ import java.util.Map;
 import org.eclipse.jdt.core.formatter.DefaultCodeFormatterConstants;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.internal.compiler.util.Util;
+import org.eclipse.jdt.neon.formatter.DefaultCodeFormatterOptions;
 
 /**
  * This is still subject to changes before 3.0.
